+++ Application/Bootstrap/Bootstrap.c
@@ -26,6 +26,7 @@ WITHOUT WARRANTIES OR REPRESENTATIONS OF ANY KIND, EITHER EXPRESS OR IMPLIED.
 #include <Library/MemoryAllocationLib.h>
 #include <Library/OcDevicePathLib.h>
 #include <Library/OcFileLib.h>
+#include <Library/OcStringLib.h>
 #include <Library/UefiBootServicesTableLib.h>
 #include <Library/UefiRuntimeServicesTableLib.h>
 #include <Library/UefiApplicationEntryPoint.h>
@@ -34,21 +35,83 @@ STATIC
 EFI_STATUS
 LoadOpenCore (
   IN  EFI_SIMPLE_FILE_SYSTEM_PROTOCOL  *FileSystem,
+  IN  EFI_DEVICE_PATH_PROTOCOL         *LoaderDevicePath,
   IN  EFI_HANDLE                       ParentImageHandle,
-  OUT EFI_HANDLE                       *ImageHandle
+  OUT EFI_HANDLE                       *ImageHandle,
+  OUT EFI_DEVICE_PATH_PROTOCOL         **ImagePath
   )
 {
   EFI_STATUS                 Status;
   VOID                       *Buffer;
+  UINTN                      LoaderPathSize;
+  UINTN                      PrefixPathSize;
+  UINTN                      RootPathLength;
+  CHAR16                     *LoaderPath;
   UINT32                     BufferSize;
 
   ASSERT (FileSystem != NULL);
   ASSERT (ParentImageHandle != NULL);
   ASSERT (ImageHandle != NULL);
+  ASSERT (ImagePath != NULL);
 
+  Buffer = NULL;
   BufferSize = 0;
-  Buffer = ReadFile (FileSystem, OPEN_CORE_DRIVER_PATH, &BufferSize, BASE_16MB);
+
+  LoaderPath = OcCopyDevicePathFullName (LoaderDevicePath);
+  LoaderPathSize = StrSize (LoaderPath);
+
+  *ImagePath = NULL;
+
+  //
+  // Try relative path: EFI\\XXX\\Bootstrap\\Bootstrap.efi -> EFI\\XXX\\OpenCore.efi
+  //
+  if (LoaderPath != NULL) {
+    if (UnicodeGetParentDirectory (LoaderPath)
+      && UnicodeGetParentDirectory (LoaderPath)) {
+      DEBUG ((DEBUG_INFO, "BS: Relative path - %s\n", LoaderPath));
+      PrefixPathSize = StrSize (LoaderPath);
+      if (LoaderPathSize - PrefixPathSize >= L_STR_SIZE (OPEN_CORE_DRIVER_PATH)) {
+        RootPathLength = PrefixPathSize / sizeof (CHAR16) - 1;
+        LoaderPath[RootPathLength] = '\\';
+        CopyMem (&LoaderPath[RootPathLength + 1], OPEN_CORE_DRIVER_PATH, L_STR_SIZE (OPEN_CORE_DRIVER_PATH));
+        Buffer = ReadFile (FileSystem, LoaderPath, &BufferSize, BASE_16MB);
+        DEBUG ((DEBUG_INFO, "BS: Startup path - %s (%p)\n", LoaderPath, Buffer));
+
+        if (Buffer != NULL) {
+          *ImagePath = FileDevicePath (NULL, LoaderPath);
+          if (*ImagePath == NULL) {
+            DEBUG ((DEBUG_INFO, "BS: File DP allocation failure, aborting\n"));
+            FreePool (Buffer);
+            Buffer = NULL;
+          }
+        }
+      }
+    }
+
+    FreePool (LoaderPath);
+  }
+
+  //
+  // Try absolute path: EFI\\BOOT\\BOOTx64.efi -> EFI\\OC\\OpenCore.efi
+  //
   if (Buffer == NULL) {
+    Buffer = ReadFile (
+      FileSystem,
+      OPEN_CORE_ROOT_PATH L"\\" OPEN_CORE_DRIVER_PATH,
+      &BufferSize,
+      BASE_16MB
+      );
+    if (Buffer != NULL) {
+      //
+      // Failure to allocate this one is not too critical, as we will still be able
+      // to choose it as a default path.
+      //
+      *ImagePath = FileDevicePath (NULL, OPEN_CORE_ROOT_PATH L"\\" OPEN_CORE_DRIVER_PATH);
+    }
+  }
+
+  if (Buffer == NULL) {
+    ASSERT (*ImagePath == NULL);
     DEBUG ((DEBUG_ERROR, "BS: Failed to locate valid OpenCore image - %p!\n", Buffer));
     return EFI_NOT_FOUND;
   }
@@ -70,6 +133,9 @@ LoadOpenCore (
   if (EFI_ERROR (Status)) {
     DEBUG ((DEBUG_ERROR, "BS: Failed to load OpenCore image - %r\n", Status));
     FreePool (Buffer);
+    if (*ImagePath != NULL) {
+      FreePool (*ImagePath);
+    }
     return Status;
   }
 
@@ -84,6 +150,9 @@ LoadOpenCore (
   if (EFI_ERROR (Status)) {
     DEBUG ((DEBUG_WARN, "BS: Failed to start OpenCore image - %r\n", Status));
     gBS->UnloadImage (*ImageHandle);
+    if (*ImagePath != NULL) {
+      FreePool (*ImagePath);
+    }
   }
 
   FreePool (Buffer);
@@ -125,6 +194,9 @@ StartOpenCore (
     return EFI_UNSUPPORTED;
   }
 
+  //
+  // For NULL path it will provide the path for the partition itself.
+  //
   AbsPath = AbsoluteDevicePath (LoadHandle, LoadPath);
 
   Status = Bootstrap->ReRun (Bootstrap, FileSystem, AbsPath);
@@ -147,7 +219,7 @@ UefiMain (
   EFI_LOADED_IMAGE_PROTOCOL         *LoadedImage;
   EFI_SIMPLE_FILE_SYSTEM_PROTOCOL   *FileSystem;
   EFI_HANDLE                        OcImageHandle;
-
+  EFI_DEVICE_PATH_PROTOCOL          *OcImagePath;
 
   DEBUG ((DEBUG_INFO, "BS: Starting OpenCore...\n"));
 
@@ -189,20 +261,24 @@ UefiMain (
   //
 
   DEBUG ((DEBUG_INFO, "BS: Trying to start loaded OpenCore image...\n"));
-  Status = StartOpenCore (FileSystem, LoadedImage->DeviceHandle, LoadedImage->FilePath);
+  Status = StartOpenCore (FileSystem, LoadedImage->DeviceHandle, NULL);
   if (EFI_ERROR (Status) && Status != EFI_NOT_FOUND) {
     return Status;
   }
 
   DEBUG ((DEBUG_INFO, "BS: Trying to load OpenCore image...\n"));
-  Status = LoadOpenCore (FileSystem, ImageHandle, &OcImageHandle);
+  Status = LoadOpenCore (FileSystem, LoadedImage->FilePath, ImageHandle, &OcImageHandle, &OcImagePath);
   if (EFI_ERROR (Status)) {
     DEBUG ((DEBUG_WARN, "BS: Failed to load OpenCore from disk - %r\n", Status));
     return EFI_NOT_FOUND;
   }
 
-  Status = StartOpenCore (FileSystem, LoadedImage->DeviceHandle, LoadedImage->FilePath);
+  Status = StartOpenCore (FileSystem, LoadedImage->DeviceHandle, OcImagePath);
   DEBUG ((DEBUG_WARN, "BS: Failed to start OpenCore image - %r\n", Status));
+
+  if (OcImagePath != NULL) {
+    FreePool (OcImagePath);
+  }
 
   return Status;
 }