+++ Include/Acidanthera/Library/OcDevicePathLib.h
@@ -187,6 +187,19 @@ OcFileDevicePathFullName (
   );
 
 /**
+  Retrieves full file path from device path allocating it on pool.
+
+  @param[in]  DevicePath  Device path to extract file path from.
+
+  @returns Path allocated from pool.
+  @retval NULL when missing.
+**/
+CHAR16 *
+OcCopyDevicePathFullName (
+  IN EFI_DEVICE_PATH_PROTOCOL        *DevicePath
+  );
+
+/**
   Duplicate device path with DevicePathInstance appended if it is not present.
 
   @param[in] DevicePath          Device Path to append new instance to, optional.