+++ Library/OcBootManagementLib/PolicyManagement.c
@@ -281,9 +281,7 @@ OcGetBootDevicePathType (
   OUT BOOLEAN                   *IsGeneric  OPTIONAL
   )
 {
-  EFI_DEVICE_PATH_PROTOCOL    *CurrNode;
   CHAR16                      *Path;
-  UINTN                       PathSize;
   UINTN                       PathLen;
   UINTN                       RestLen;
   UINTN                       Index;
@@ -298,31 +296,12 @@ OcGetBootDevicePathType (
     *IsFolder = FALSE;
   }
 
-  for (CurrNode = DevicePath; !IsDevicePathEnd (CurrNode); CurrNode = NextDevicePathNode (CurrNode)) {
-    if ((DevicePathType (CurrNode) == MEDIA_DEVICE_PATH)
-     && (DevicePathSubType (CurrNode) == MEDIA_FILEPATH_DP)) {
-      //
-      // Perform copying of all the underlying nodes due to potential unaligned access.
-      //
-      PathSize = OcFileDevicePathFullNameSize (CurrNode);
-      if (PathSize == 0) {
-        return OC_BOOT_UNKNOWN;
-      }
-
-      Path = AllocatePool (PathSize);
-      if (Path == NULL) {
-        return OC_BOOT_UNKNOWN;
-      }
-
-      OcFileDevicePathFullName (Path, (FILEPATH_DEVICE_PATH *) CurrNode, PathSize);
-      PathLen = StrLen (Path);
-      break;
-    }
-  }
-
+  Path = OcCopyDevicePathFullName (DevicePath);
   if (Path == NULL) {
     return OC_BOOT_UNKNOWN;
   }
+
+  PathLen = StrLen (Path);
 
   //
   // Use the trailing character to determine folder.